[UB FIX] Fix Division by Zero and Signed Integer Overflows
BUG TYPE: Undefined Behavior / Logic Bug
SEVERITY: HIGH
ISSUE:
1. Division by Zero in `IOMapOTBM::loadMap` when loading 0-byte files.
2. Signed Integer Overflow (UB) in `LiveSocket` and `LiveClient` when packing coordinates.
3. Signed Integer Overflow (UB) in `LightDrawer::draw` when calculating squared distances on large maps.

FIX:
1. Added `f.size() > 0` check in `source/io/iomap_otbm.cpp`.
2. Added `static_cast<uint32_t>` casts in `source/live/live_socket.cpp` and `source/live/live_client.cpp`.
3. Used `int64_t` for distance calculations in `source/rendering/utilities/light_drawer.cpp`.

+++ source/live/live_client.cpp
@@ -323,8 +323,8 @@ void LiveClient::sendReady() {
 
 void LiveClient::queryNode(int32_t ndx, int32_t ndy, bool underground) {
 	uint32_t nd = 0;
-	nd |= ((ndx >> 2) << 18);
-	nd |= ((ndy >> 2) << 4);
+	nd |= (static_cast<uint32_t>(ndx >> 2) << 18);
+	nd |= (static_cast<uint32_t>(ndy >> 2) << 4);
 	nd |= (underground ? 1 : 0);
 	queryNodeList.insert(nd);
 }
